Moved position class to folder
